Cli, export services

+++ src/main/java/com/news/storage/ArticleTagLinker.java
@@ -3,5 +3,5 @@ package com.news.storage;
 import java.sql.SQLException;
 
 public interface ArticleTagLinker {
-    void linkArticleTags(int articleId, int tagId) throws SQLException;
+    void linkArticleTags(long articleId, int tagId) throws SQLException;
 }
